Modification of sample data
Just reduced the number of specimens included in the sample data so that those who wish to run and/or troubleshoot these scripts can do so in a shorter period of time. I have kept 100 specimens in the latest sample data so that the script should run to completion in a minute or less (depending on the number of cores allocated).